requirements.txt
scipy
tensorboardX
scikit-image
Pillow
opencv-python==4.7.0.72
timm==0.6.13
packaging
pyyaml 
requests
tqdm
ipython
opencv-python==4.7.0.72
thop
matplotlib
numba==0.56.4
tabulate
mmcv-full==1.7.1
pandas
tensorboard